chore: added raylib as a git submodule
Previously, our Makefile would clone raylib's git repository as part of
the build process. Our approach worked, but it was a little awkward
(e.g. the repository was removed after invoking `make clean`).

A more elegant approach is simply adding raylib as a git submodule and
letting git deal with it! Not only does this streamline our build
process, but it also locks builds to a specific version of raylib.

If you want to learn more about submodules, here is a great article
about them:

https://github.blog/2016-02-01-working-with-submodules/