refactor: clean up `EntityManager` and `EventManager`
Cleaned up the way we were managing entities and events in `scene`. We
were managing entities and events by essentially using a stack with
confusing variable names, and a make-shift stack api in the form of
a bunch of `static` functions in `scene.c`. Addressed this using the
new `deque` implementation.