refactor: replace `QuadtreeCreate` with `QuadtreeNew`
We were having issues with the topmost Quadtree being freed (since it
always lived on the stack). In order to address this we decided to just
make Quadtree's constructor return a `Quadtree*` (i.e. a pointer to heap
allocated memory).

It's worth noting that we now have a new convention: add `New` to the
end of a method if it returns a pointer to heap allocated memory.

Co-authored-by: Austin Aurelio <austin@delveam.com>