feat: smooth out one-way platform collision
Before this commit, the player could awkwardly resolve itself to the top
of a one-way platform (despite not jumping high enough to reach it).

The following changes make it so, in similar situations, the player can
only resolve itself a maximum of 3 pixels. In practice, this behavior
looks much more natural.