refactor: remove `TAG_COLOR` from all entity's dependencies
CColor is currently just used for tinting textures (which we don't even
do in the first place...).

Every current entity that required TAG_COLOR could get away with just
hard-coding a color in its draw method. I wanted to just remove CColor
outright, but I ultimately decided to just leave it there in case we
need it in the future (e.g. the colorful Firework entity).