refactor: add a layer of indirection to `CPlayer`
The new SceneDefer approach allows us to access an array separate from
the ECS that contains the real Player data; CPlayer just contains a
handle to said array.

This is somewhat against the ECS approach, but CPlayer was getting too
big and also I can do whatever I want to!